Toute l'ihm

+++ Equipe.java
@@ -163,8 +163,8 @@ public class Equipe implements Util {
      * @return Une note représentant la performance collective de l'équipe.
      */
     @Override 
-    public int participer() {
-        int res = 0;
+    public double participer() {
+        double res = 0;
         for (Athlete athlete : lesAthletes) {
             res += athlete.getAgilite() + athlete.getForce() + athlete.getEndurance();
         }
