+++ Util.java
@@ -1,5 +1,5 @@
 public interface Util{
     public int getNbVictoire();
     public void setNbVictoire(int nbVictoire);
-    public int participer();
+    public double participer();
 }
